+++ HISTORY.rst
@@ -6,6 +6,7 @@ Unreleased
 ----------
 
 * The `SUPERFLUOUS_NO_LINT` warning has been removed from the glinter. It likely did more harm than good, and makes it hard to make `metrics.yaml` files that pass across different versions of `glean_parser`.
+* Expired metrics will now produce a linter warning, `EXPIRED_METRIC`.
 
 1.28.3 (2020-07-28)
 -------------------